+++ src/test/java/com/evanzeimet/wirecucumber/WireCucumberTest.java
@@ -1,9 +1,6 @@
 package com.evanzeimet.wirecucumber;
 
 import static org.junit.Assert.assertNull;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.notNull;
-import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
@@ -49,16 +46,4 @@ public class WireCucumberTest {
 		assertNull(actualException);
 	}
 
-	@Test
-	public void initialize_nullOptions() {
-		WireCucumberOptions givenOptions = null;
-
-		doNothing().when(wireCucumber).startWireMockServer(any());
-		doNothing().when(wireCucumber).createSteps(any());
-
-		wireCucumber.initialize(givenOptions);
-
-		verify(wireCucumber).startWireMockServer(notNull());
-		verify(wireCucumber).createSteps(notNull());
-	}
 }